requirements.txt
# Main project dependencies
azure-storage-blob==12.16.0
python-dotenv==1.0.0
azure-eventhub==5.7.0
azure-identity==1.12.0
azure-functions==1.14.0
opencv-python-headless==4.7.0.72
pytest==7.3.1
pytest-asyncio==0.21.0
pydantic==1.10
aiohttp